Move extension expressions to their own file
